tests: move the tests to a better place
